pyproject.toml
[tool.poetry]
name = "cryptography-basis"
version = "0.1.0"
description = ""
authors = ["Paw565pl <paw565pl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
